Implement fork to always require auxpow blocks.
See https://forum.namecoin.info/viewtopic.php?f=5&t=2466.

In particular, this commit implements a hard fork active for all blocks
with a nTime later than 2020-01-01 (to be changed once a point has been
fixed in discussions).  All blocks afterwards no longer use nVersion to
encode an auxpow flag and the chain ID.  Instead, all blocks must
necessarily be merge mined, thus auxpow flag is assumed to be true.  The
chain ID is instead stored into the nNonce field, which is not used
otherwise for merge-mined blocks.

All existing tests (unit and regression) have been updated as necessary
to succeed with the changes, but there are not yet any tests added that
exercise the behaviour after the fork point.  This will be added in
future commits.  Also to be added are other things included in the same
fork, like a removal of the existing BDB lock limit.