Introduce "transition window" for same-chain ID parent blocks.
Add a second fork time a day after the initial always-auxpow fork and
require that parent blocks must not have the same chain ID until that
later point in time.

If we would not do that, it would be possible during the transition to
use a parent block with PoW and nTime before the fork as Namecoin block
and, at the same time, as auxpow for a block with nTime after the fork.
While this is presumably not a terribly bad thing (as it is temporary
and not economic to use for a miner), it should not happen.

With the new logic, this is no longer possible:  After the additional
day is over and arbitrary parent blocks are accepted as auxpow, the
network won't accept blocks with nTime before the fork anymore due to
the existing restrictions (nTime must be within node time and,
furthermore, after the MedianTimePast).