Remove strict checks for Namecoin tx nVersion at fork.
At the always-auxpow fork, no longer require transaction nVersion to be
set to the "magic" Namecoin value.  Instead, the nVersion will be
ignored for tx validation.  This is possible since all information is
actually present in whether or not the outputs contain Namecoin scripts.
This helps in the future if upstream Bitcoin wants to change nVersion
(e. g., opt-in RBF or malleability fixes).

The patch removes the strict check when validating blocks, but still
enforces the policy on the mempool and creates transactions with the
version set by default.  However, with this in, we can in the future
change that (after the fork) without requiring any forking change.