Disable failing rpc_psbt.py test assertion.
This disables an assertion in the rpc_psbt.py test that has been failing
for Namecoin since that test was introduced.  At least on the 0.17
release branch, we should have clean test runs with no known failures
polluting the result.

This might be related to https://github.com/bitcoin/bitcoin/issues/13738.